+++ package.json
@@ -40,6 +40,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.3",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "^5.0.1"
   }
 }
