+++ src/index.html
@@ -14,6 +14,6 @@
     <title>Top Games</title>
   </head>
   <body>
-    <div id="app-container"></div>
+    <div id="root"></div>
   </body>
 </html>
