+++ src/scripts/Layout/Root.module.scss
@@ -1,4 +1,4 @@
-.app {
+.root {
   min-height: 100vh;
   display: grid;
   grid-template-rows: auto 1fr;
